+++ Source/OnlineSubsystemAccelByte/Private/OnlineIdentityInterfaceAccelByte.cpp
@@ -548,7 +548,11 @@ void FOnlineIdentityAccelByte::GetLinkedAccountAuthToken(int32 LocalUserNum, con
 	Delegate.ExecuteIfBound(LocalUserNum, bWasSuccessful, AuthToken);
 }
 
+#if ENGINE_MAJOR_VERSION == 5 && ENGINE_MINOR_VERSION >= 4
+void FOnlineIdentityAccelByte::GetUserPrivilege(const FUniqueNetId& UserId, EUserPrivileges::Type Privilege, const FOnGetUserPrivilegeCompleteDelegate& Delegate, EShowPrivilegeResolveUI ShowResolveUI)
+#else
 void FOnlineIdentityAccelByte::GetUserPrivilege(const FUniqueNetId& UserId, EUserPrivileges::Type Privilege, const FOnGetUserPrivilegeCompleteDelegate& Delegate)
+#endif
 {
 	AB_OSS_PTR_INTERFACE_TRACE_BEGIN(TEXT("UserId: %s"), *UserId.ToDebugString());